Add DB and dependency injection functionality
This commit creates the database structure with the entity, DAO,
repositories and the database itself.
It also implements the dependency injection by Dagger Hilt and
creates the TaskApp and AppModule files.
It creates the firt util files (OrderType and TaskOrder).
Finally, it implements the firt use cases (GetTasks and DeleteTask)
and creates the TaskUseCases file.